Safari-611.1.10/LayoutTests/js/dfg-uint32array-overflow-constant-expected.txt

Tests that storing a value that is outside of the int32 range into a Uint32Array results in correct wrap-around.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS array[0] is 0x8005465c
PASS successfullyParsed is true

TEST COMPLETE
